Initial implementation of Azure auth support for DRS
Add support for using Microsoft Azure default credentials
when running in a Terra Azure Interactive Analysis Cloud Environment.
These credentials are used to auth with Terra backend services,
including Martha/terra-drs-hub and Rawls.

When running in a Terra Google IA Cloud Environment,
the behavior is as before, with no changes.

This initial implementation is thought to be functionally complete,
yet tests and documentation remain to be added.